5.4. MVVM - Model-view-viewmodel/Scott Lilly/12/MVVM-tutorial/MVVM-tutorial/MVVM_PatternVersion/ViewModel/AccountCreationViewModel.cs
using MVVM_tutorial.MVVM_PatternVersion.Model;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;


namespace MVVM_tutorial.MVVM_PatternVersion.ViewModel
{
    public class AccountCreationViewModel
    {
        public AccountModel NewAccount { get; set; }

        public AccountCreationViewModel()
        {
            NewAccount = new AccountModel();
        }

        public void ValidatePassword()
        {
            NewAccount.ValidatePassword();
        }
    }
}
